+++ tests/20-main.spec.js
@@ -4,16 +4,19 @@
 import {
   createList, decodeList, createCredential, checkStatus, statusTypeMatches,
   assertStatusList2021Context, getCredentialStatus
-} from '..';
+} from '../lib/index.js';
 import * as didKey from '@digitalbazaar/did-method-key';
-import {extendContextLoader} from 'jsonld-signatures';
-import {slCredential as SLC} from './mock-sl-credential.js';
+import jsigs from 'jsonld-signatures';
+import {
+  slCredentialRevocation as SLCRevocation,
+  slCredentialSuspension as SLCSuspension
+} from './mock-sl-credentials.js';
 import statusListCtx from '@digitalbazaar/vc-status-list-context';
-import vc from '@digitalbazaar/vc';
+import {defaultDocumentLoader} from '@digitalbazaar/vc';
 import {Ed25519Signature2020} from '@digitalbazaar/ed25519-signature-2020';
 import suiteCtx2020 from 'ed25519-signature-2020-context';
 
-const {defaultDocumentLoader} = vc;
+const {extendContextLoader} = jsigs;
 
 const VC_SL_CONTEXT_URL = statusListCtx.constants.CONTEXT_URL_V1;
 const VC_SL_CONTEXT = statusListCtx.contexts.get(VC_SL_CONTEXT_URL);
@@ -26,7 +29,8 @@ const encodedList100k =
 const documents = new Map();
 documents.set(VC_SL_CONTEXT_URL, VC_SL_CONTEXT);
 documents.set(SUITE_CONTEXT_URL, SUITE_CONTEXT);
-documents.set(SLC.id, SLC);
+documents.set(SLCRevocation.id, SLCRevocation);
+documents.set(SLCSuspension.id, SLCSuspension);
 
 const didKeyDriver = didKey.driver();
 
@@ -89,7 +93,8 @@ describe('createCredential', () => {
   it('should create a StatusList2021Credential credential', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     credential.should.be.an('object');
     credential.should.deep.equal({
       '@context': [
@@ -101,7 +106,8 @@ describe('createCredential', () => {
       credentialSubject: {
         id: `${id}#list`,
         type: 'StatusList2021',
-        encodedList: encodedList100k
+        encodedList: encodedList100k,
+        statusPurpose: 'revocation'
       }
     });
   });
@@ -124,9 +130,9 @@ describe('statusTypeMatches', () => {
         id: 'https://example.com/status/1#67342',
         type: 'StatusList2021Entry',
         statusListIndex: '67342',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const result = statusTypeMatches({credential});
     result.should.equal(true);
@@ -148,9 +154,9 @@ describe('statusTypeMatches', () => {
         id: 'https://example.com/status/1#67342',
         type: 'ex:NotMatch',
         statusListIndex: '67342',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const result = statusTypeMatches({credential});
     result.should.equal(false);
@@ -173,7 +179,8 @@ describe('statusTypeMatches', () => {
   it('should fail when "@context" is not an array', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     let err;
     let result;
     try {
@@ -192,7 +199,8 @@ describe('statusTypeMatches', () => {
   it('should fail when first "@context" value is unexpected', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     let err;
     let result;
     try {
@@ -211,7 +219,8 @@ describe('statusTypeMatches', () => {
   it('should fail when "credentialStatus" does not exist', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     let err;
     let result;
     try {
@@ -229,7 +238,8 @@ describe('statusTypeMatches', () => {
     '"statusTypeMatches"', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     let err;
     let result;
     try {
@@ -249,7 +259,8 @@ describe('statusTypeMatches', () => {
     '"@context"', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     let err;
     let result;
     try {
@@ -258,7 +269,7 @@ describe('statusTypeMatches', () => {
         id: 'https://example.com/status/1#50000',
         type: 'StatusList2021Entry',
         statusListIndex: '50000',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       };
       result = statusTypeMatches({credential});
     } catch(e) {
@@ -287,9 +298,9 @@ describe('checkStatus', () => {
         type: 'StatusList2021Entry',
         statusPurpose: 'revocation',
         statusListIndex: '67342',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const suite = new Ed25519Signature2020();
     const result = await checkStatus({
@@ -320,9 +331,9 @@ describe('checkStatus', () => {
         type: 'StatusList2021Entry',
         statusPurpose: 'revocation',
         statusListIndex: '67342',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const suite = new Ed25519Signature2020();
     const result = await checkStatus({
@@ -335,9 +346,9 @@ describe('checkStatus', () => {
   });
 
   it('should fail to verify an invalid status list vc', async () => {
-    const invalidSLC = JSON.parse(JSON.stringify(SLC));
+    const invalidSLC = JSON.parse(JSON.stringify(SLCRevocation));
     delete invalidSLC.proof;
-    invalidSLC.id = 'https://example.com/status/no-proof-slc';
+    invalidSLC.id = 'https://example.com/status/no-proof-SLCRevocation';
     documents.set(invalidSLC.id, invalidSLC);
     const credential = {
       '@context': [
@@ -374,10 +385,10 @@ describe('checkStatus', () => {
 
   it('should fail to verify status list vc that has been tampered with',
     async () => {
-      const invalidSLC = JSON.parse(JSON.stringify(SLC));
+      const invalidSLC = JSON.parse(JSON.stringify(SLCRevocation));
       // intentionally change it's type
       invalidSLC.type = ['VerifiableCredential', 'ex:Invalid'];
-      invalidSLC.id = 'https://example.com/status/tampered-slc';
+      invalidSLC.id = 'https://example.com/status/tampered-SLCRevocation';
       documents.set(invalidSLC.id, invalidSLC);
       const credential = {
         '@context': [
@@ -413,9 +424,9 @@ describe('checkStatus', () => {
 
   it('should verify with an invalid status list vc when ' +
     '"verifyStatusListCredential" is set to "false"', async () => {
-    const invalidSLC = JSON.parse(JSON.stringify(SLC));
+    const invalidSLC = JSON.parse(JSON.stringify(SLCRevocation));
     delete invalidSLC.proof;
-    invalidSLC.id = 'https://example.com/status/no-proof-invalid-slc';
+    invalidSLC.id = 'https://example.com/status/no-proof-invalid-SLCRevocation';
     documents.set(invalidSLC.id, invalidSLC);
     const credential = {
       '@context': [
@@ -463,9 +474,9 @@ describe('checkStatus', () => {
         type: 'StatusList2021Entry',
         statusPurpose: 'revocation',
         statusListIndex: '67342',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const suite = new Ed25519Signature2020();
     const result = await checkStatus({
@@ -476,6 +487,44 @@ describe('checkStatus', () => {
     });
     should.not.exist(result.error);
     result.verified.should.equal(true);
+  });
+
+  it('should fail to verify if status purpose in credential does not match ' +
+    'the status purpose of status list credential', async () => {
+    const credential = {
+      '@context': [
+        'https://www.w3.org/2018/credentials/v1',
+        VC_SL_CONTEXT_URL
+      ],
+      id: 'urn:uuid:a0418a78-7924-11ea-8a23-10bf48838a41',
+      type: ['VerifiableCredential', 'example:TestCredential'],
+      credentialSubject: {
+        id: 'urn:uuid:4886029a-7925-11ea-9274-10bf48838a41',
+        'example:test': 'foo'
+      },
+      credentialStatus: {
+        id: 'https://example.com/status/2#67342',
+        type: 'StatusList2021Entry',
+        statusPurpose: 'suspension',
+        statusListIndex: '67342',
+        // intentionally point the statusListCredential to the
+        // status list credential with status purpose "revocation".
+        statusListCredential: SLCRevocation.id
+      },
+      issuer: SLCRevocation.issuer,
+    };
+    const suite = new Ed25519Signature2020();
+    const result = await checkStatus({
+      credential,
+      suite,
+      documentLoader,
+      verifyStatusListCredential: true
+    });
+    should.exist(result.error);
+    result.error.message.should.equal(
+      'The status purpose "revocation" of the status list credential does ' +
+      'not match the status purpose "suspension" in the credential.');
+    result.verified.should.equal(false);
   });
 
   it('should verify multiple statuses of a credential', async () => {
@@ -495,15 +544,15 @@ describe('checkStatus', () => {
         type: 'StatusList2021Entry',
         statusPurpose: 'revocation',
         statusListIndex: '67342',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       }, {
-        id: 'https://example.com/status/1#67343',
+        id: 'https://example.com/status/2#67343',
         type: 'StatusList2021Entry',
         statusPurpose: 'suspension',
         statusListIndex: '67343',
-        statusListCredential: SLC.id
+        statusListCredential: SLCSuspension.id
       }],
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const suite = new Ed25519Signature2020();
     const result = await checkStatus({
@@ -532,9 +581,9 @@ describe('checkStatus', () => {
         id: 'https://example.com/status/1#67342',
         type: 'ex:NonmatchingStatusType',
         statusListIndex: '67342',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const suite = new Ed25519Signature2020();
     const result = await checkStatus({
@@ -566,15 +615,15 @@ describe('checkStatus', () => {
         type: 'StatusList2021Entry',
         statusPurpose: 'revocation',
         statusListIndex: '67342',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       }, {
         id: 'https://example.com/status/1#67342',
         type: 'ex:NonmatchingStatusType',
         statusPurpose: 'revocation',
         statusListIndex: '67342',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       }],
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const suite = new Ed25519Signature2020();
     const result = await checkStatus({
@@ -603,9 +652,9 @@ describe('checkStatus', () => {
         id: 'https://example.com/status/1#67342',
         type: 'StatusList2021Entry',
         statusPurpose: 'revocation',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const suite = new Ed25519Signature2020();
     const result = await checkStatus({
@@ -637,7 +686,7 @@ describe('checkStatus', () => {
         statusPurpose: 'suspension',
         statusListIndex: '67342'
       },
-      issuer: SLC.issuer,
+      issuer: SLCSuspension.issuer,
     };
     const suite = new Ed25519Signature2020();
     const result = await checkStatus({
@@ -668,9 +717,9 @@ describe('checkStatus', () => {
         id: 'https://example.com/status/1#50000',
         type: 'StatusList2021Entry',
         statusListIndex: '50000',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const suite = new Ed25519Signature2020();
     const result = await checkStatus({
@@ -705,9 +754,9 @@ describe('checkStatus', () => {
         statusListIndex: '50000',
         // intentionally set statusListCredential to an id that is not set
         // in documents
-        statusListCredential: 'https://example.com/status/2'
+        statusListCredential: 'https://example.com/status/3'
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const suite = new Ed25519Signature2020();
     const result = await checkStatus({
@@ -720,15 +769,15 @@ describe('checkStatus', () => {
     should.exist(result.error);
     result.error.message.should.equal('Could not load ' +
       '"StatusList2021Credential"; reason: Document loader unable to load ' +
-      'URL "https://example.com/status/2".');
+      'URL "https://example.com/status/3".');
   });
 
   it('should fail when "statusListCredential" type does not ' +
     'include "StatusList2021Credential"', async () => {
-    const invalidSLC = JSON.parse(JSON.stringify(SLC));
-    // intentionally set SLC type to an invalid type
+    const invalidSLC = JSON.parse(JSON.stringify(SLCRevocation));
+    // intentionally set SLCRevocation type to an invalid type
     invalidSLC.type = ['InvalidType'];
-    invalidSLC.id = 'https://example.com/status/invalid-slc-type';
+    invalidSLC.id = 'https://example.com/status/invalid-SLCRevocation-type';
 
     documents.set(invalidSLC.id, invalidSLC);
 
@@ -750,7 +799,7 @@ describe('checkStatus', () => {
         statusListIndex: '50000',
         statusListCredential: invalidSLC.id
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const result = await checkStatus({
       credential, documentLoader, verifyStatusListCredential: false
@@ -763,7 +812,7 @@ describe('checkStatus', () => {
 
   it('should fail when "credentialSubject" type is not ' +
     '"StatusList2021"', async () => {
-    const invalidSLC = JSON.parse(JSON.stringify(SLC));
+    const invalidSLC = JSON.parse(JSON.stringify(SLCRevocation));
     // intentionally set credential subject type to an invalid type
     invalidSLC.credentialSubject.type = 'InvalidType';
     invalidSLC.id = 'https://example.com/status/invalid-sl-type';
@@ -788,7 +837,7 @@ describe('checkStatus', () => {
         statusListIndex: '50000',
         statusListCredential: invalidSLC.id
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const result = await checkStatus({
       credential, documentLoader, verifyStatusListCredential: false
@@ -801,7 +850,7 @@ describe('checkStatus', () => {
 
   it('should fail when "credentialSubject.encodedList" ' +
     'cannot not be decoded', async () => {
-    const invalidSLC = JSON.parse(JSON.stringify(SLC));
+    const invalidSLC = JSON.parse(JSON.stringify(SLCRevocation));
     // intentionally set encodedList to an invalid value
     invalidSLC.credentialSubject.encodedList = 'INVALID-XYZ';
     invalidSLC.id = 'https://example.com/status/invalid-encoded-list';
@@ -826,7 +875,7 @@ describe('checkStatus', () => {
         statusListIndex: '50000',
         statusListCredential: invalidSLC.id
       },
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
     };
     const result = await checkStatus({
       credential, documentLoader, verifyStatusListCredential: false
@@ -869,7 +918,7 @@ describe('checkStatus', () => {
       credentialStatus: {
         id: 'https://example.com/status/1#67342',
         type: 'StatusList2021Entry',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       }
     };
     const documentLoader = 'https://example.com/status/1';
@@ -909,7 +958,7 @@ describe('checkStatus', () => {
         id: 'https://example.com/status/1#50000',
         type: 'StatusList2021Entry',
         statusListIndex: '50000',
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       }
     };
     const documentLoader = extendContextLoader(async url => {
@@ -953,7 +1002,7 @@ describe('checkStatus', () => {
         VC_SL_CONTEXT_URL
       ],
       id: 'urn:uuid:e74fb1d6-7926-11ea-8e11-10bf48838a41',
-      issuer: SLC.issuer,
+      issuer: SLCRevocation.issuer,
       issuanceDate: '2021-03-10T04:24:12.164Z',
       type: ['VerifiableCredential', 'example:TestCredential'],
       credentialSubject: {
@@ -965,7 +1014,7 @@ describe('checkStatus', () => {
         type: 'StatusList2021Entry',
         statusPurpose: 'revocation',
         statusListIndex: 50000,
-        statusListCredential: SLC.id
+        statusListCredential: SLCRevocation.id
       }
     };
     let err;
@@ -1009,7 +1058,7 @@ describe('checkStatus', () => {
         type: 'StatusList2021Entry',
         statusPurpose: 'revocation',
         statusListIndex: '67342',
-        statusListCredential: SLC.id,
+        statusListCredential: SLCRevocation.id,
       },
       // this issuer does not match the issuer for the mock SLC specified
       // by `SLC.id` above
@@ -1047,7 +1096,7 @@ describe('checkStatus', () => {
         type: 'StatusList2021Entry',
         statusPurpose: 'revocation',
         statusListIndex: '67342',
-        statusListCredential: SLC.id,
+        statusListCredential: SLCRevocation.id,
       },
       // this issuer does not match the issuer for the mock SLC specified
       // by `SLC.id` above
@@ -1086,7 +1135,8 @@ describe('assertStatusList2021Context', () => {
   it('should fail when "@context" is not an array', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     let err;
     let result;
     try {
@@ -1105,7 +1155,8 @@ describe('assertStatusList2021Context', () => {
   it('should fail when first "@context" value is unexpected', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     let err;
     let result;
     try {
@@ -1124,7 +1175,8 @@ describe('assertStatusList2021Context', () => {
   it('should fail when "CONTEXTS.RL_V1" is not in "@context"', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     let err;
     let result;
     try {
@@ -1158,7 +1210,8 @@ describe('getCredentialStatus', () => {
   it('should fail when "credentialStatus" is not an object', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     let err;
     let result;
     try {
@@ -1177,13 +1230,14 @@ describe('getCredentialStatus', () => {
     '"StatusList2021Entry"', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     credential.credentialStatus = {
       id: 'https://example.com/status/1#67342',
       type: 'InvalidType',
       statusPurpose: 'revocation',
       statusListIndex: '67342',
-      statusListCredential: SLC.id
+      statusListCredential: SLCRevocation.id
     };
     let err;
     let result;
@@ -1203,20 +1257,21 @@ describe('getCredentialStatus', () => {
     'with correct type', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     credential.credentialStatus = [{
-      id: 'https://example.com/status/1#67342',
+      id: 'https://example.com/status/2#67342',
       type: 'ex:NonmatchingStatusType',
       statusPurpose: 'suspension',
       statusListIndex: '67342',
-      statusListCredential: SLC.id
+      statusListCredential: SLCSuspension.id
     },
     {
       id: 'https://example.com/status/1#67342',
       type: 'StatusList2021Entry',
       statusPurpose: 'revocation',
       statusListIndex: '67342',
-      statusListCredential: SLC.id
+      statusListCredential: SLCRevocation.id
     }];
     let err;
     let result;
@@ -1234,7 +1289,8 @@ describe('getCredentialStatus', () => {
     'array', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     credential.credentialStatus = [ ];
     let err;
     let result;
@@ -1253,20 +1309,21 @@ describe('getCredentialStatus', () => {
     'matching "StatusList2021Entry"', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     credential.credentialStatus = [{
       id: 'https://example.com/status/1#12345',
       type: 'ex:NonmatchingStatusType',
       statusPurpose: 'revocation',
       statusListIndex: '12345',
-      statusListCredential: SLC.id
+      statusListCredential: SLCRevocation.id
     },
     {
       id: 'https://example.com/status/1#67342',
       type: 'ex:NonmatchingStatusType',
       statusPurpose: 'suspension',
       statusListIndex: '67342',
-      statusListCredential: SLC.id
+      statusListCredential: SLCSuspension.id
     }];
     let err;
     let result;
@@ -1285,13 +1342,14 @@ describe('getCredentialStatus', () => {
     '"StatusList2021Entry" and "statusPurpose" matches', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     credential.credentialStatus = {
       id: 'https://example.com/status/1#67342',
       type: 'StatusList2021Entry',
       statusPurpose: 'revocation',
       statusListIndex: '67342',
-      statusListCredential: SLC.id
+      statusListCredential: SLCRevocation.id
     };
     let err;
     let result;
@@ -1308,13 +1366,14 @@ describe('getCredentialStatus', () => {
   it('should fail when "statusPurpose" is not specified', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     credential.credentialStatus = {
       id: 'https://example.com/status/1#67342',
       type: 'StatusList2021Entry',
       statusPurpose: 'revocation',
       statusListIndex: '67342',
-      statusListCredential: SLC.id
+      statusListCredential: SLCRevocation.id
     };
     let err;
     let result;
@@ -1333,13 +1392,14 @@ describe('getCredentialStatus', () => {
     '"credentialStatus.statusPurpose"', async () => {
     const id = 'https://example.com/status/1';
     const list = await createList({length: 100000});
-    const credential = await createCredential({id, list});
+    const credential = await createCredential(
+      {id, list, statusPurpose: 'revocation'});
     credential.credentialStatus = {
       id: 'https://example.com/status/1#67342',
       type: 'StatusList2021Entry',
       statusPurpose: 'revocation',
       statusListIndex: '67342',
-      statusListCredential: SLC.id
+      statusListCredential: SLCRevocation.id
     };
     let err;
     let result;