+++ CHANGELOG.md
@@ -1,4 +1,24 @@
-# vc-status-list ChangeLog
+# @digitalbazaar/vc-status-list ChangeLog
+
+## 5.0.0 - 2022-06-16
+
+### Changed
+- **BREAKING**: Convert to module (ESM).
+- **BREAKING**: Require Node.js >=14.
+- Update dependencies.
+- Lint module.
+
+## 4.0.0 - 2022-06-04
+
+### Added
+- **BREAKING**: Add required param `statusPurpose` to `createCredential()`.
+- Check if `statusPurpose` in credential matches the `statusPurpose` of
+  status list credential. If they don't match, an error will be thrown.
+
+## 3.1.0 - 2022-05-30
+
+### Added
+- Export `StatusList` utility class.
 
 ## 3.0.0 - 2022-04-15